Empty drafts at one place have one drawing, not two
Tree and StartLine each wired NewRow for parked ghosts and the live
caret. That is one concept — empty drafts standing at an anchor — split
across two files, so a blur-slot or resume change had to land twice.
Ghosts is that drawing, and both consumers use it.
